+++ package.json
@@ -27,7 +27,8 @@
   "peerDependencies": {
     "@angular/core": ">=4.3.6",
     "@angular/http": ">=4.3.6",
-    "@skyux/assets": "^3.0.0"
+    "@skyux/assets": "^3.0.0",
+    "@skyux/core": "^3.2.2"
   },
   "dependencies": {},
   "devDependencies": {
