+++ src/app/public/modules/i18n/lib-resources.service.ts
@@ -10,7 +10,7 @@ import {
 
 import {
   SkyI18nFormat
-} from './format';
+} from '../../utils/format';
 
 import {
   SkyAppLocaleInfo
