+++ src/app/public/modules/i18n/resources.service.spec.ts
@@ -27,7 +27,7 @@ import {
 
 import {
   SkyI18nFormat
-} from './format';
+} from '../../utils/format';
 
 import {
   SkyAppLocaleProvider
