--- src/app/public/modules/i18n/window.service.ts
@@ -1,10 +1,0 @@
-import {
-  Injectable
-} from '@angular/core';
-
-@Injectable()
-export class SkyI18nWindowService {
-  public get nativeWindow(): Window {
-    return window;
-  }
-}
