+++ src/app/public/utils/format.ts
@@ -1,3 +1,5 @@
+// Using this class instead of the `SkyAppFormat` found in `@skyux/core`
+// to avoid a circular dependency.
 export class SkyI18nFormat {
   public static formatText(format: string, ...args: any[]): string {
     return String(format).replace(
